+++ src/components/base/i2c_target_communication_with_obc.hpp
@@ -6,7 +6,7 @@
 #ifndef S2E_COMPONENTS_BASE_I2C_TARGET_COMMUNICATION_WITH_OBC_HPP_
 #define S2E_COMPONENTS_BASE_I2C_TARGET_COMMUNICATION_WITH_OBC_HPP_
 
-#include "../../interface/hils/hils_port_manager.hpp"
+#include "../../simulation/hils/hils_port_manager.hpp"
 #include "../cdh/obc.hpp"
 #include "uart_communication_with_obc.hpp"
