+++ src/simulation/CMakeLists.txt
@@ -20,6 +20,8 @@ add_library(${PROJECT_NAME} STATIC
   ground_station/ground_station.cpp
   
   inter_spacecraft_communication/inter_spacecraft_communication.cpp
+
+  hils/hils_port_manager.cpp
 )
 
 include(../../common.cmake)
